+++ wallet/bindings/wasm/lib/bindings.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-// This file overwrites the `bindings.ts` file from `bindings/nodejs/lib`, to link the Wasm `MessageHandler` interface.
+// This file overwrites the `bindings.ts` file from `wallet/bindings/nodejs/lib`, to link the Wasm `MessageHandler` interface.
 // The rest of the TypeScript definitions are copied as-is to the `out` directory before being compiled.
 
 // Import needs to be in a single line, otherwise it breaks